# document/openapi.yml
openapi: '3.0.2'
info:
  title: WalletAPI Sample
  description: |
    # WalletAPI
    - auth
    - users
    - nft
  # termsOfService: 'https://example.com/'
  # contact:
  #   name: "Support"
  #   url: "会社のURL等"
  #   email: "sample@email.com"
  version: '1.0'
servers:
  - url: 'http://localhost:{port}'
    description: 'Local Development'
    variables:
      port:
        enum: ['8000', '8080']
        default: '8000'
...
paths:
...
  '/users/user-info/{userId}':
    get:
      tags:
        - users
      summary: ユーザー情報を取得する
      description: 指定したユーザーの情報をDBから取得する
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/userInfo'
        '401':
          description: 認証されていないアクセス
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '404':
          description: 指定のリソースが存在しない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
      security:
        - Bearer: []
  '/users/wallet-address/{userId}':
    get:
      tags:
        - users
      summary: ウォレットアドレスを取得する
      description: 指定したユーザーのウォレットアドレスをDBから取得する
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/userWalletAddress'
        '401':
          description: 認証されていないアクセス
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '404':
          description: 指定のリソースが存在しない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
      security:
        - Bearer: []
  '/users/update-pass/{userId}':
    patch:
      tags:
        - users
      summary: パスワードを更新する
      description: 指定したユーザーのパスワードを更新する
      requestBody:
        description: パスワード更新用リクエストボディ
        required: true
        content:
          application/json:
            schema:
              type: 'object'
              properties:
              $ref: '#/components/schemas/updatePassword'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: 'object'
                properties:
                  user:
                    type: string
                    example: 'success'
        '401':
          description: 認証されていないアクセス
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '404':
          description: 指定のリソースが存在しない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
      security:
        - Bearer: []
...
components:
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer
      description: Credentials or access token for API
  schemas:
...
    Unauthorized:
      title: 認証されてないリクエスト
      type: object
      properties:
        code:
          description: |-
            認証の失敗です。
            ログインするとき、入力した内容に誤りがあるとこのエラーを返します。
          type: string
          example: 401
        message:
          description: 画面に表示するエラーメッセージ
          type: string
          example: 認証されていないアクセスです。
    NotFound:
      title: 指定のリソースが存在しない
      type: object
      properties:
        code:
          description: |-
            リソースが見つかりません。
            指定された条件のデータが存在しないときにこのエラーを返します。
          type: string
          example: 404
        message:
          description: 画面に表示するエラーメッセージ
          type: string
          example: 指定のリソースは存在しません。
...
    updatePassword:
      type: object
      properties:
        password: { type: string }
...
    userInfo:
      type: object
      properties:
        userId: { type: number }
        email: { type: string }
        wallet_address: { type: string }
    userWalletAddress:
      type: object
      properties:
        wallet_address: { type: string }
...